fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -23,12 +23,12 @@ void printFibonacciNumbers(int n)
     printf("%d, ", f1);
     for (i = 1; i < n; i++) {
         printf("%d, ", f2);
-        int next = f1 + f2;
+        long int next = f1 + f2;
         f1 = f2;
         f2 = next;
     }
 }
-int main()
+long int main()
 {
     printFibonacciNumbers(50);
     return 0;
